+++ src/ref_soft/r_bsp.c
@@ -84,8 +84,8 @@ void R_RotateBmodel (void)
 // yaw
 	angle = currententity->angles[YAW];		
 	angle = angle * M_TWOPI / 360;
-	s = sin(angle);
-	c = cos(angle);
+	s = (float)sin(angle);
+	c = (float)cos(angle);
 
 	temp1[0][0] = c;
 	temp1[0][1] = s;
@@ -101,8 +101,8 @@ void R_RotateBmodel (void)
 // pitch
 	angle = currententity->angles[PITCH];		
 	angle = angle * M_TWOPI / 360;
-	s = sin(angle);
-	c = cos(angle);
+	s = (float)sin(angle);
+	c = (float)cos(angle);
 
 	temp2[0][0] = c;
 	temp2[0][1] = 0;
@@ -119,8 +119,8 @@ void R_RotateBmodel (void)
 // roll
 	angle = currententity->angles[ROLL];		
 	angle = angle * M_TWOPI / 360;
-	s = sin(angle);
-	c = cos(angle);
+	s = (float)sin(angle);
+	c = (float)cos(angle);
 
 	temp1[0][0] = 1;
 	temp1[0][1] = 0;
