+++ src/ref_soft/r_poly.c
@@ -768,14 +768,14 @@ void R_PolygonScanLeftEdge (void)
 	if (lmaxindex == 0)
 		lmaxindex = r_polydesc.nump;
 
-	vtop = ceil (r_polydesc.pverts[i].v);
+	vtop = (float)ceil(r_polydesc.pverts[i].v);
 
 	do
 	{
 		pvert = &r_polydesc.pverts[i];
 		pnext = pvert - 1;
 
-		vbottom = ceil (pnext->v);
+		vbottom = (float)ceil(pnext->v);
 
 		if (vtop < vbottom)
 		{
@@ -831,7 +831,7 @@ void R_PolygonScanRightEdge (void)
 	if (vvert > r_refdef.fvrectbottom_adj)
 		vvert = r_refdef.fvrectbottom_adj;
 
-	vtop = ceil (vvert);
+	vtop = (float)ceil(vvert);
 
 	do
 	{
@@ -844,7 +844,7 @@ void R_PolygonScanRightEdge (void)
 		if (vnext > r_refdef.fvrectbottom_adj)
 			vnext = r_refdef.fvrectbottom_adj;
 
-		vbottom = ceil (vnext);
+		vbottom = (float)ceil(vnext);
 
 		if (vtop < vbottom)
 		{
@@ -1163,8 +1163,8 @@ static void R_DrawPoly( int iswater )
 		pverts++;
 	}
 
-	ymin = ceil (ymin);
-	ymax = ceil (ymax);
+	ymin = (float)ceil(ymin);
+	ymax = (float)ceil(ymax);
 
 	if (ymin >= ymax)
 		return;		// doesn't cross any scans at all
